Move DrawTarget and iter tests back into their mods
