+++ core/src/iterator/mod.rs
@@ -80,3 +80,32 @@ where
         pixel::Translate::new(self, offset)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use embedded_graphics::{
+        geometry::Point, iterator::PixelIteratorExt, mock_display::MockDisplay,
+        pixelcolor::BinaryColor, Pixel,
+    };
+
+    #[test]
+    fn draw_pixel_iterator() {
+        let pixels = [
+            Pixel(Point::new(0, 0), BinaryColor::On),
+            Pixel(Point::new(1, 0), BinaryColor::Off),
+            Pixel(Point::new(2, 0), BinaryColor::On),
+            Pixel(Point::new(2, 1), BinaryColor::Off),
+        ];
+
+        let mut display = MockDisplay::new();
+        pixels.iter().copied().draw(&mut display).unwrap();
+
+        assert_eq!(
+            display,
+            MockDisplay::from_pattern(&[
+                "#.#", //
+                "  .", //
+            ])
+        );
+    }
+}
